pyproject.toml
[tool.poetry]
authors = ["The SKA SDP Continuum Imaging Pipeline Authors"]
description = "SKA SDP Continuum Imaging Pipeline"
documentation = "https://developer.skao.int/projects/ska-sdp-continuum-imaging-pipeline"
license = "BSD-3-Clause"
name = "ska_sdp_cip"
readme = "README.md"
repository = "https://gitlab.com/ska-telescope/sdp/ska-sdp-continuum-imaging-pipeline"
version = "0.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.scripts]
ska-sdp-cip = "ska_sdp_cip.apps.pipeline_app:main"

[tool.poetry.dependencies]
# Must specify an upper bound for Python version, otherwise some dependencies
# cannot technically meet the requirement.
python = "^3.10,<3.13"
numpy = "^1.23"
python-casacore = "^3.5.2"
ducc0 = "^0.34"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
build = "^0.7.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pylint = "^2.12.2"
pylint-junit = "^0.3.2"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

# Install with `poetry install --with docs`
[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.0"
sphinx-rtd-theme = "^1.0.0"

[tool.black]
line-length = 79
